refactor(ContentClientConfig): split into seperate files for each interface
