+++ src/lib/config/CommonContentClientConfig.ts
@@ -1,19 +1,5 @@
 import { AxiosAdapter } from 'axios';
 
-export interface ContentClientConfigV1 extends CommonContentClientConfig {
-  /**
-   * Account to retrieve content from.
-   */
-  account: string;
-}
-
-export interface ContentClientConfigV2 extends CommonContentClientConfig {
-  /**
-   * Hub name to retrieve content from.
-   */
-  hubName: string;
-}
-
 /**
  * Configuration settings for Content Delivery API client. You can optionally
  * override these values with environment specific values.
@@ -31,7 +17,7 @@ export interface CommonContentClientConfig {
   stagingEnvironment?: string;
 
   /**
-   * If set, the SDK will request content using the locale settings provided.
+   * If set, the SDK will request content using the locale settings provided.rm -fr
    * If your content contains any localized fields, this will cause a single
    * locale to be returned rather than the complete list of values.
    */
