+++ src/lib/media/Media.ts
@@ -1,4 +1,4 @@
-import { CommonContentClientConfig } from '../ContentClientConfig';
+import { CommonContentClientConfig } from '../config/CommonContentClientConfig';
 import { ImageMeta, VideoMeta } from './MediaMeta';
 
 /**
